elf: _dl_fixup*: make @reloc_arg 64-bit safe
This came up during arc64 glibc bringup on qemu usermode.

On ARC, lazy rezolver calls _dl_fixup() with 2nd argument as PC of the
PLT entry beiing fixed up. Thus this needs to be full 64-bit value for
arc64, however _dl_fixup() defines @reloc_arg as ElfW(Word) which maps
to Elf64_Word -> uint32_t causing gcc to truncate the value to 32-bits.

This problem show on 64-bit ARC and not other arches likely due to the
semantics of @reloc_arg. On ARC it is the runtime PC of PLT entry whereas for
others it might be the index or something similar which doesn't need
more than 32-bits.

Reported-by: Cupertino Miranda <cmiranda@synopsys.com>
Reported-by: Vladimir Isaev <isaev@synopsys.com>
Link: https://github.com/foss-for-synopsys-dwc-arc-processors/toolchain/issues/403
Signed-off-by: Vineet Gupta <vgupta@synopsys.com>